Update: (WE-10763) Forgot to rename the HOC feature
