+++ packages/es-components/src/components/controls/buttons/OutlineButton.js
@@ -3,7 +3,7 @@ import PropTypes from 'prop-types';
 import styled from 'styled-components';
 import tinycolor from 'tinycolor2';
 import { useTheme } from '../../util/useTheme';
-import { withLoadingStateWhileRunning } from './features/withDisabledLoadingStateContent';
+import { withLoadingStateWhileRunning } from './features/withLoadingStateWhileRunning';
 
 const StyledButton = styled.button`
   background-color: ${props => props.colors.bgColor};
